+++ src/App.js
@@ -1,9 +1,9 @@
 import React, { useState, useEffect } from 'react';
 import { Routes, Route } from 'react-router-dom';
-import { Navigate } from 'react-router-dom';
 import { getAllData } from './util/index';
 import { HomePage } from './components/HomePage/HomePage';
 import { LoginPage } from './components/LoginPage/LoginPage';
+import ProtectedRoute from './components/ProtectedRoute/ProtectedRoute';
 
 const URL = 'http://localhost:8000/api/v1/';
 
@@ -24,32 +24,19 @@ const [message, setMessage] = useState('');
 
   }, []);
 
-  const ProtectedRoute = ({ children }) => {
-    // Placeholder to actually check if user is authenticated
-    const useAuth = () => {
-      if (Math.random() > 0.5) {
-        return true;
-      }
-      return false;
-    }
-
-    const user = useAuth();
-    if (!user) {
-      // User is not authenticated
-      return <Navigate to="/login" />;
-    }
-    return children;
-  };
-
   return (
     <>
       <h1>{message}</h1>
-      <ProtectedRoute>
-        <Routes>
-          <Route exact path="/" element={<HomePage />} />
-        </Routes>  
-      </ProtectedRoute> 
+      
       <Routes>
+        <Route 
+          exact path="/" 
+          element={
+            <ProtectedRoute>
+              <HomePage />
+            </ProtectedRoute>
+          } 
+        />
         <Route  path="/login" element={<LoginPage />} />
       </Routes> 
     </>
